+++ tests/testthat/test-semmcci-mc-latent-med-defined-none.R
@@ -19,10 +19,32 @@ lapply(
       data = data
     )
     set.seed(seed)
-    results <- MC(
+    results_null <- MC(
       fit,
       R = R,
-      alpha = c(0.001, 0.01, 0.05)
+      alpha = c(0.001, 0.01, 0.05),
+      decomposition = NULL
+    )
+    set.seed(seed)
+    results_chol <- MC(
+      fit,
+      R = R,
+      alpha = c(0.001, 0.01, 0.05),
+      decomposition = "chol"
+    )
+    set.seed(seed)
+    results_eigen <- MC(
+      fit,
+      R = R,
+      alpha = c(0.001, 0.01, 0.05),
+      decomposition = "eigen"
+    )
+    set.seed(seed)
+    results_svd <- MC(
+      fit,
+      R = R,
+      alpha = c(0.001, 0.01, 0.05),
+      decomposition = "svd"
     )
     set.seed(seed)
     answers <- MASS::mvrnorm(
@@ -31,22 +53,67 @@ lapply(
       Sigma = lavaan::vcov(fit)
     )
     testthat::test_that(
-      text,
+      paste(text, "NULL"),
       {
         testthat::expect_equal(
-          results$thetahat$est[names(lavaan::coef(fit))],
+          results_null$thetahat$est[names(lavaan::coef(fit))],
           as.vector(lavaan::coef(fit)),
           check.attributes = FALSE
         )
         testthat::expect_true(
           abs(
-            .MCCI(results)["visual~~textual", "0.05%"] - quantile(answers[, "visual~~textual"], .0005)
+            .MCCI(results_null)["visual~~textual", "97.5%"] - quantile(answers[, "visual~~textual"], .975, na.rm = TRUE)
+          ) <= tol
+        )
+      }
+    )
+    testthat::test_that(
+      paste(text, "chol"),
+      {
+        testthat::expect_equal(
+          results_chol$thetahat$est[names(lavaan::coef(fit))],
+          as.vector(lavaan::coef(fit)),
+          check.attributes = FALSE
+        )
+        testthat::expect_true(
+          abs(
+            .MCCI(results_chol)["visual~~textual", "97.5%"] - quantile(answers[, "visual~~textual"], .975, na.rm = TRUE)
+          ) <= tol
+        )
+      }
+    )
+    testthat::test_that(
+      paste(text, "eigen"),
+      {
+        testthat::expect_equal(
+          results_eigen$thetahat$est[names(lavaan::coef(fit))],
+          as.vector(lavaan::coef(fit)),
+          check.attributes = FALSE
+        )
+        testthat::expect_true(
+          abs(
+            .MCCI(results_eigen)["visual~~textual", "97.5%"] - quantile(answers[, "visual~~textual"], .975, na.rm = TRUE)
+          ) <= tol
+        )
+      }
+    )
+    testthat::test_that(
+      paste(text, "svd"),
+      {
+        testthat::expect_equal(
+          results_svd$thetahat$est[names(lavaan::coef(fit))],
+          as.vector(lavaan::coef(fit)),
+          check.attributes = FALSE
+        )
+        testthat::expect_true(
+          abs(
+            .MCCI(results_svd)["visual~~textual", "97.5%"] - quantile(answers[, "visual~~textual"], .975, na.rm = TRUE)
           ) <= tol
         )
       }
     )
   },
-  R = 1000L,
+  R = 5000L,
   tol = 0.05,
   text = "test-semmcci-mc-latent-med-defined-none"
 )