+++ tests/testthat/test-semmcci-mc-simple-med-defined-none.R
@@ -29,10 +29,31 @@ lapply(
       fixed.x = FALSE
     )
     set.seed(seed)
-    results <- MC(
+    results_null <- MC(
       fit,
       R = R,
-      alpha = c(0.001, 0.01, 0.05)
+      alpha = c(0.001, 0.01, 0.05),
+      decomposition = NULL
+    )
+    set.seed(seed)
+    results_chol <- MC(
+      fit,
+      R = R,
+      alpha = c(0.001, 0.01, 0.05),
+      decomposition = "chol"
+    )
+    set.seed(seed)
+    results_eigen <- MC(
+      fit,
+      R = R,
+      alpha = c(0.001, 0.01, 0.05),
+      decomposition = "eigen"
+    )
+    results_svd <- MC(
+      fit,
+      R = R,
+      alpha = c(0.001, 0.01, 0.05),
+      decomposition = "svd"
     )
     set.seed(seed)
     answers <- MASS::mvrnorm(
@@ -44,23 +65,68 @@ lapply(
       answers
     )
     testthat::test_that(
-      text,
+      paste(text, "NULL"),
       {
         testthat::expect_equal(
-          results$thetahat$est,
+          results_null$thetahat$est,
           lavaan::parameterEstimates(fit)$est,
           check.attributes = FALSE
         )
         testthat::expect_true(
           abs(
-            .MCCI(results)["cp", "0.05%"] - quantile(answers[, "cp"], .0005)
+            .MCCI(results_null)["cp", "97.5%"] - quantile(answers[, "cp"], .975, na.rm = TRUE)
+          ) <= tol
+        )
+      }
+    )
+    testthat::test_that(
+      paste(text, "chol"),
+      {
+        testthat::expect_equal(
+          results_chol$thetahat$est,
+          lavaan::parameterEstimates(fit)$est,
+          check.attributes = FALSE
+        )
+        testthat::expect_true(
+          abs(
+            .MCCI(results_chol)["cp", "97.5%"] - quantile(answers[, "cp"], .975, na.rm = TRUE)
+          ) <= tol
+        )
+      }
+    )
+    testthat::test_that(
+      paste(text, "eigen"),
+      {
+        testthat::expect_equal(
+          results_eigen$thetahat$est,
+          lavaan::parameterEstimates(fit)$est,
+          check.attributes = FALSE
+        )
+        testthat::expect_true(
+          abs(
+            .MCCI(results_eigen)["cp", "97.5%"] - quantile(answers[, "cp"], .975)
+          ) <= tol
+        )
+      }
+    )
+    testthat::test_that(
+      paste(text, "svd"),
+      {
+        testthat::expect_equal(
+          results_svd$thetahat$est,
+          lavaan::parameterEstimates(fit)$est,
+          check.attributes = FALSE
+        )
+        testthat::expect_true(
+          abs(
+            .MCCI(results_svd)["cp", "97.5%"] - quantile(answers[, "cp"], .975, na.rm = TRUE)
           ) <= tol
         )
       }
     )
   },
   n = 1000L,
-  R = 1000L,
+  R = 5000L,
   tol = 0.05,
   text = "test-semmcci-mc-simple-med-defined-none"
 )